fix: changed paraid in genesis

+++ scripts/update_hardcoded_specs.py
@@ -26,7 +26,7 @@ SPECS = [
     "integritee-kusama",
     "integritee-polkadot",
     "integritee-moonbase",
-    # "shell-kusama-lease2", # enable if you want to change the data for registration.
+    "shell-kusama", # enable if you want to change the data for registration.
 ]
 COLLATOR = "target/release/integritee-collator"
 RES_DIR = "polkadot-parachains/chain-specs"
@@ -46,29 +46,33 @@ def main(regenesis: bool):
         orig_file = f'{RES_DIR}/{chain_spec}.json'
         new_file_base = f'{RES_DIR}/{chain_spec}-fresh'
 
-        with open(orig_file, 'r+') as spec_orig_file:
-            orig_json = json.load(spec_orig_file)
+        if not os.path.exists(orig_file):
+            print(f"no previous spec found for {orig_file}, not migrating previous settings")
+            os.popen(f"cp {new_file_base}-raw.json {orig_file}")
+        else:
+            with open(orig_file, 'r+') as spec_orig_file:
+                orig_json = json.load(spec_orig_file)
 
-            # migrate old values to new spec
-            with open(f'{new_file_base}-raw.json', 'r+') as spec_new_file:
-                new_json = json.load(spec_new_file)
+                # migrate old values to new spec
+                with open(f'{new_file_base}-raw.json', 'r+') as spec_new_file:
+                    new_json = json.load(spec_new_file)
 
-                new_json["bootNodes"] = orig_json["bootNodes"]
+                    new_json["bootNodes"] = orig_json["bootNodes"]
 
-                if not regenesis:
-                    new_json["genesis"] = orig_json["genesis"]
+                    if not regenesis:
+                        new_json["genesis"] = orig_json["genesis"]
 
-                # go to beginning of the file to overwrite
-                spec_orig_file.seek(0)
-                json.dump(new_json, spec_orig_file, indent=2)
-                spec_orig_file.truncate()
+                    # go to beginning of the file to overwrite
+                    spec_orig_file.seek(0)
+                    json.dump(new_json, spec_orig_file, indent=2)
+                    spec_orig_file.truncate()
 
-        # remove side-products
-        os.remove(f'{new_file_base}.json')
-        os.remove(f'{new_file_base}-raw.json')
-        os.remove(f'{new_file_base}-raw-unsorted.json')
-        os.remove(f'{new_file_base}.state')
-        os.remove(f'{new_file_base}.wasm')
+            # remove side-products
+            os.remove(f'{new_file_base}.json')
+            os.remove(f'{new_file_base}-raw.json')
+            os.remove(f'{new_file_base}-raw-unsorted.json')
+            os.remove(f'{new_file_base}.state')
+            os.remove(f'{new_file_base}.wasm')
 
 
 if __name__ == '__main__':
